requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
autopep8==2.3.2
aws-opentelemetry-distro==0.10.1
aws-requests-auth==0.4.3
beautifulsoup4==4.13.5
bedrock-agentcore==0.1.3
bedrock-agentcore-starter-toolkit==0.1.10
blinker==1.9.0
boto3==1.40.32
botocore==1.40.32
cachetools==5.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
ddgs==5.5.5
Deprecated==1.2.18
dill==0.3.8
docstring_parser==0.17.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
googleapis-common-protos==1.70.0
grpcio==1.75.0
h11==0.16.0
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==8.6.1
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
lxml==6.0.1
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mcp==1.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
marshmallow==2.5.0
numpy==2.3.3
# ... (rest truncated for brevity) ...
streamlit==1.49.1
pandas==2.3.2
